Move test trigger stuff to sim files
